[BetterReflectionDocBlock] fix FixedSerializerTest
